I re-implemented isTouchingCeiling testing but better...
The old isTouchingCeiling test relied on raycasting from a checkceiling object that was a child of the player.
The problem with that approach was that it was relying on a single point (the center position of that checkceiling object). Our player is a square and it may be the case that its center point is clear (nothing above) but one of its corners is blocked above.
So now we raycast from 2 or 3 points. The topmost 2 or 3 vertices of the player (these change when the player rotates).
We check 2 when the player is parallel to the ground (that is, his z angle is a multiple of 90 degrees).
We check 3 when the player is at an angle to the ground (we ignore its bottommost vertex).